# services/catalog/openapi.yaml
openapi: 3.0.3
info:
  title: catalog-service
  version: 0.1.0
  description: >
    Carreras, cursos, tomos, lecciones, bancos de preguntas y precios.
    Invariante I-5: respuesta_correcta y solucion_esperada NUNCA aparecen en
    ninguna respuesta pública.
servers:
  - url: /api/catalog
paths:
...
  /cursos/{slugOId}:
    get:
      summary: Detalle de un curso publicado con su estructura (público)
      parameters:
        - { name: slugOId, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Curso con tomos y lecciones (sin contenido) }
        '404': { description: No existe o no está publicado }
...
  /lecciones/{id}:
    get:
      summary: Contenido de una lección (autenticado; gating de matrícula A-18)
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Bloques y ejercicios (sin solucion_esperada — I-5) }
        '401': { description: Sin token }
  /tomos/{id}/evaluacion:
    get:
      summary: Evaluación del tomo (autenticado, sin respuesta_correcta — I-5)
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Preguntas sin respuestas }
  /interno/bancos/{id}/respuestas:
    get:
      summary: Respuestas de un banco (INTERNO, solo enrollment con x-interno-token — A-19)
      parameters:
        - { name: x-interno-token, in: header, required: true, schema: { type: string } }
      responses:
        '200': { description: Preguntas con respuestaCorrecta, puntaje y nivel }
        '401': { description: Token interno inválido }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }